Prompt/requirements.txt
fastapi==0.103.2
langchain==0.0.308
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.23.2
